% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_histogram.R
\name{auto_histogram}
\alias{auto_histogram}
\title{Auto Histogram}
\usage{
auto_histogram(df)
}
\arguments{
\item{df}{a dataframe}
}
\value{
a ggplot matrix of histograms
}
\description{
This function takes a dataframe and returns a matrix of ggplot
geom_histograms of all the numeric columns.
}
