% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_boxgrid.R
\name{bb_boxgrid}
\alias{bb_boxgrid}
\title{bb_boxgrid}
\usage{
bb_boxgrid(df, title = NULL)
}
\arguments{
\item{df}{a dataframe}

\item{title}{a title}
}
\value{
a ggplot grid of boxplots
}
\description{
Takes a dataframe and returns a matrix of ggplot geom_boxplots of all numeric columns.
}
