% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_boxmatrix.R
\name{bb_boxmatrix}
\alias{bb_boxmatrix}
\title{bb_boxmatrix}
\usage{
bb_boxmatrix(df, title = NULL)
}
\arguments{
\item{df}{a dataframe}

\item{title}{a title}
}
\value{
a ggplot matrix of boxplots
}
\description{
Takes a dataframe and returns a matrix of ggplot geom_boxplots of all numeric columns.
}
