% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_cleancols.R
\name{bb_cleancols}
\alias{bb_cleancols}
\title{bb_cleancols}
\usage{
bb_cleancols(df)
}
\arguments{
\item{df}{a dataframe}

\item{title}{a title}
}
\value{
a df with nice column labels
}
\description{
Takes a dataframe and cleans up column labels
}
