% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_corrplot.R
\name{bb_corrplot}
\alias{bb_corrplot}
\title{bb_corrplot takes a dataframe and makes a nice correlation plot
using corrr}
\usage{
bb_corrplot(df)
}
\arguments{
\item{df}{a dataframe}
}
\value{
a kable-formatted table
}
\description{
bb_corrplot takes a dataframe and makes a nice correlation plot
using corrr
}
