% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_histgrid.R
\name{bb_histgrid}
\alias{bb_histgrid}
\title{bb_histgrid}
\usage{
bb_histgrid(df, title = NULL)
}
\arguments{
\item{df}{a dataframe}

\item{title}{a title}
}
\value{
a ggplot grid of histograms
}
\description{
Takes a dataframe and returns a grid of histograms of all the numeric columns.
}
