% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_regression_metrics.R
\name{bb_regression_metrics}
\alias{bb_regression_metrics}
\title{bb_regression_metrics}
\usage{
bb_regression_metrics(pred, resp, colname = "metrics")
}
\arguments{
\item{pred}{a vector of predicted values}

\item{colname}{an optional column label}

\item{obs}{a vector of observed values}
}
\value{
a df
}
\description{
Takes vectors of predictions and observations from regression models, returns common error metrics in a df using yardstick.
}
