% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_scattergrid_y.R
\name{bb_scattergrid_y}
\alias{bb_scattergrid_y}
\title{bb_scattergrid_y}
\usage{
bb_scattergrid_y(df, title = NULL)
}
\arguments{
\item{df}{a dataframe}

\item{title}{a title}
}
\value{
a ggplot grid of scatterplots, with a line of best fit (lm)
}
\description{
Takes a dataframe with a response variable in the first column, returns a grid of scatterplots of each predictor's relationship to the response, plus lm line of best fit.
}
