% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_screeplot.R
\name{bb_screeplot}
\alias{bb_screeplot}
\title{bb_screeplot
Principal Component Analysis - PCA}
\usage{
bb_screeplot(prcomp, plot_type = "bar", title = NULL)
}
\arguments{
\item{prcomp}{a stats::prcomp object}

\item{plot_type}{'bar' or 'line', bar is default}

\item{title}{a title}
}
\value{
a ggplot barplot or lineplot
}
\description{
This function takes a prcomp object and returns a simple scree plot
using ggplot. Specify 'bar' or 'line' as second parameter.
}
