% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_table.R
\name{bb_table}
\alias{bb_table}
\title{bb_table takes a dataframe and makes a nice table using kable.}
\usage{
bb_table(df, cap = NULL, cols = NULL, dig = 3, fw = F)
}
\arguments{
\item{df}{a dataframe}

\item{cap}{a caption}

\item{cols}{a vector of column names to include}

\item{dig}{number of decimal places for rounding}

\item{fw}{a boolean T/F for full-width}
}
\value{
a kable-formatted table
}
\description{
bb_table takes a dataframe and makes a nice table using kable.
}
