% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_screeplot.R
\name{gg_screeplot}
\alias{gg_screeplot}
\title{gg_screeplot
Principal Component Analysis - PCA}
\usage{
gg_screeplot(prcomp, plot_type = "bar")
}
\arguments{
\item{prcomp}{a stats::prcomp object}

\item{plot_type}{'bar' or 'line', bar is default}
}
\value{
a ggplot barplot or lineplot
}
\description{
This function takes a prcomp object and returns a simple scree plot
using ggplot. Specify 'bar' or 'line' as second parameter.
}
