% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram_matrix.R
\name{histogram_matrix}
\alias{histogram_matrix}
\title{Histogram Matrix}
\usage{
histogram_matrix(df)
}
\arguments{
\item{df}{a dataframe}
}
\value{
a ggplot matrix of histograms
}
\description{
This function takes a dataframe and returns a matrix of ggplot
geom_histograms of all the numeric columns.
}
