% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scree_plot.R
\name{scree_plot}
\alias{scree_plot}
\title{Scree Plot
Principal Component Analysis - PCA}
\usage{
scree_plot(prcomp)
}
\arguments{
\item{prcomp}{a stats::prcomp object}
}
\value{
a ggplot
}
\description{
This function takes a prcomp object and returns a simple scree plot.
}
