% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_corrplot.R
\name{bb_corrplot}
\alias{bb_corrplot}
\title{bb_corrplot takes a dataframe and makes a nice correlation plot
of the numeric columns using corrr}
\usage{
bb_corrplot(df, title = NULL)
}
\arguments{
\item{df}{a dataframe}

\item{title}{a title}
}
\value{
a corrr rplot
}
\description{
bb_corrplot takes a dataframe and makes a nice correlation plot
of the numeric columns using corrr
}
