% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_histmatrix.R
\name{bb_histmatrix}
\alias{bb_histmatrix}
\title{bb_histmatrix}
\usage{
bb_histmatrix(df)
}
\arguments{
\item{df}{a dataframe}
}
\value{
a ggplot matrix of histograms
}
\description{
This function takes a dataframe and returns a matrix of ggplot
geom_histograms of all the numeric columns.
}
