% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_regression_metrics.R
\name{bb_regression_metrics}
\alias{bb_regression_metrics}
\title{bb_regression_metrics}
\usage{
bb_regression_metrics(pred, obs, dig = 3, title = NA)
}
\arguments{
\item{pred}{a vector of predicted values}

\item{obs}{a vector of observed values}

\item{dig}{optional number of decimal places for rounding - default 3}

\item{title}{an optional title}

\item{fw}{optional boolean T/F for full-width - default F}
}
\value{
a kable-formatted table
}
\description{
Takes vectors of predictions and observations from regression models, returns common error metrics in a table using yardstick.
}
